moved MapControls to own component
