+++ applications/fishing-map/src/features/map/MapControls.module.css
@@ -1,0 +1,12 @@
+.mapControls {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: absolute;
+  right: var(--space-S);
+  top: var(--space-S);
+}
+
+.mapControls > * {
+  margin-bottom: var(--space-XS);
+}
